+++ .gitignore
@@ -28,3 +28,7 @@ migrate_working_dir/
 .dart_tool/
 .packages
 build/
+
+# FVM Version Cache
+.fvm/
+.fvmrc
