+++ lib/packages/core/lib/src/convertors/convertors.dart
@@ -5,3 +5,4 @@ export 'poll/poll_converter.dart';
 export 'tag/tag_convertor.dart';
 export 'user/user_convertor.dart';
 export 'user/sdk_client_info_convertor.dart';
+export 'attachment/attachment_convertor.dart';
