+++ lib/packages/core/lib/src/utils/utils.dart
@@ -1,11 +1,20 @@
 export 'constants/strings.dart';
 export 'constants/enums.dart';
+
 export 'preferences/preferences.dart';
+
 export 'conversation/conversation_utils.dart';
+
 export 'analytics/analytics.dart';
+
 export 'notifications/notification_handler.dart';
+
 export 'callback/lm_chat_callback.dart';
 
 export 'chatroom/chatroom_utils.dart';
 export 'chatroom/explore_utils.dart';
+
 export 'extension/extension.dart';
+
+export 'media/media_handler.dart';
+export 'media/permission_handler.dart';
