+++ lib/packages/core/lib/src/views/media/configurations/forwarding/setting.dart
@@ -1,7 +1,7 @@
-/// {@template lm_chat_report_setting}
+/// {@template lm_chat_media_forwarding_setting}
 /// Setting class for the Chat MediaForwarding Screen.
 /// {@endtemplate}
 class LMChatMediaForwardingSetting {
-  /// {@macro lm_chat_report_setting}
+  /// {@macro lm_chat_media_forwarding_setting}
   const LMChatMediaForwardingSetting();
 }
