+++ lib/packages/core/lib/src/views/media/configurations/forwarding/style.dart
@@ -1,7 +1,7 @@
-/// {@template lm_chat_report_style}
+/// {@template lm_chat_media_forwarding_style}
 /// Style class for the Chat MediaForwarding Screen.
 /// {@endtemplate}
 class LMChatMediaForwardingStyle {
-  /// {@macro lm_chat_report_style}
+  /// {@macro lm_chat_media_forwarding_style}
   const LMChatMediaForwardingStyle();
 }
