+++ lib/packages/core/lib/src/widgets/widgets.dart
@@ -10,3 +10,4 @@ export 'lists/conversation/dm_conversation_list.dart';
 
 export 'explore/explore_tile.dart';
 export 'explore/explore_join_button.dart';
+export 'tagging/tagging_textfield_ta.dart';
